Drop RCS ids in upstream zlib source
We're not maintaining a this as a fork, it's upstream source with a handful
of patches. Thus, the RCS ids aren't particularly useful or important. They
are a bit of a maintenance burden and generate noise in diffs.

ok kn, no objection millert, "kill" guess who

(cherry picked from commit 46c3d7fc63d60116d0370b8e39af3255c97f6276)