Add support for the PCIe controller on the Qualcomm SC8280XP.  Thankfully UEFI already initializes those, so we can simply just make use of that.
That said, the ctrl/dbi region isn't the first in the register list, so
instead try and look it up first and use it if available.  Furthermore,
the ATU region isn't part of the ctrl/dbi region, so if we are able to
retrieve a separate reg for the ATU, use that instead.  Some reshuffling
is necessary to make that work.

Tested on my Lenovo x13s and the MacchiatoBin
ok kettenis@

(cherry picked from commit a848a3cb9519d5b1af313d52b0cefadefdc7fa6b)