Two small tweaks to the geofeed code
Only allocate b64 when it is needed. This way we can avoid allocating extra
memory for the signed data itself. Also, only check for the end signature
marker when it is actually expected. It's not forbidden - if stupid - to
have a comment '# End Signature:' in the signed data.

ok job

(cherry picked from commit f3520e92de980b5db99406519b782d922754e551)