Turn sowriteable(), sballoc() and sbfree() macro to inline functions.
soreadable() is already presented as inline function, but corresponding
sowriteable() is still macro. Also it's no reason to keep sballoc() and
sbfree() as macro.

The first argument of sballoc() and sbfree() is not used, but keep it for
a while.

ok kn@ bluhm@

(cherry picked from commit d452b926f142c4271bac299168cd77e2ced42046)