nc: do not test the connection in non-interactive mode
The connection test writes four X to the socket, which corrupts data
that we may want to pipe into nc. So don't do that if stdin is not a
tty but still do it in scan mode, this is needed according to chris.

based on a diff by and ok mpf

(cherry picked from commit 42efad6ca3ee4fa561e93c2065fa3a45fbafde35)